pyproject.toml
[project]
name = "agent-deployment-readiness-lab"
version = "0.1.0"
description = "A focused LangGraph app for turning messy workflow briefs into deployment-ready agent plans."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.0.0",
    "langchain-openai>=1.0.0",
    "langgraph>=1.0.0",
    "langgraph-cli[inmem]>=0.3.1",
    "langsmith>=0.4.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
agent-readiness = "agent_deployment_readiness_lab.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_deployment_readiness_lab"]

[tool.pytest.ini_options]
pythonpath = [
    "src",
]
testpaths = [
    "tests",
]
